+++ src/main.py
@@ -1,0 +1,11 @@
+from test.test_number_pair import test_number_to_pair
+from test.test_pair_number import test_pair_to_number
+
+
+if __name__ == '__main__':
+    test_number_to_pair(4, 'White', 'Brown')
+    test_number_to_pair(5, 'White', 'Slate')
+    test_pair_to_number('Black', 'Orange', 12)
+    test_pair_to_number('Violet', 'Slate', 25)
+    test_pair_to_number('Red', 'Orange', 7)
+    print('Done :)')
