+++ test/test_wire_color_pair.py
@@ -1,0 +1,13 @@
+from src.wire_color_pair import get_color_from_pair_number, get_pair_number_from_color
+
+
+def test_number_to_pair(pair_number,
+                        expected_major_color, expected_minor_color):
+    major_color, minor_color = get_color_from_pair_number(pair_number)
+    assert(major_color == expected_major_color)
+    assert(minor_color == expected_minor_color)
+
+
+def test_pair_to_number(major_color, minor_color, expected_pair_number):
+    pair_number = get_pair_number_from_color(major_color, minor_color)
+    assert(pair_number == expected_pair_number)
