--- test_number_pair.py
@@ -1,8 +1,0 @@
-from main import get_color_from_pair_number
-
-
-def test_number_to_pair(pair_number,
-                        expected_major_color, expected_minor_color):
-    major_color, minor_color = get_color_from_pair_number(pair_number)
-    assert(major_color == expected_major_color)
-    assert(minor_color == expected_minor_color)
